backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
django-allauth==0.63.3
celery==5.3.6
redis==5.0.6
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
yt-dlp==2023.12.30
moviepy==1.0.3
openai==1.35.3
google-generativeai==0.7.0
groq==0.9.0
requests==2.32.3
python-multipart==0.0.9
whitenoise==6.6.0
gunicorn==22.0.0
dj-database-url==2.2.0
django-storages==1.14.3
boto3==1.34.131
flake8==7.1.0
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.5.4
